update PagesActions input and output types
