+++ Apps.Webflow/Models/Request/Pages/UpdatePageContentRequest.cs
@@ -3,12 +3,14 @@ using Apps.Webflow.DataSourceHandlers.Locale;
 using Blackbird.Applications.Sdk.Common;
 using Blackbird.Applications.Sdk.Common.Dynamic;
 using Blackbird.Applications.Sdk.Common.Files;
-using Newtonsoft.Json;
 
 namespace Apps.Webflow.Models.Request.Pages;
 
 public class UpdatePageContentRequest
 {
+    [Display("Page")]
+    public FileReference File { get; set; }
+
     [Display("Page ID")]
     [DataSource(typeof(PageDataSourceHandler))]
     public string? PageId { get; set; }
@@ -16,25 +18,4 @@ public class UpdatePageContentRequest
     [Display("Locale ID")]
     [DataSource(typeof(SiteLocaleDataSourceHandler))]
     public string? LocaleId { get; set; }
-
-    [Display("HTML file")]
-    public FileReference File { get; set; }
-}
-
-public class UpdatePageDomRequest
-{
-    [JsonProperty("localeId")]
-    public string? LocaleId { get; set; }
-
-    [JsonProperty("nodes")]
-    public IEnumerable<UpdatePageNode> Nodes { get; set; }
-}
-
-public class UpdatePageNode
-{
-    [JsonProperty("nodeId")]
-    public string NodeId { get; set; }
-
-    [JsonProperty("text")]
-    public string Text { get; set; }
 }
